Moves around requires, adds app.js
Uses getFormFields for sign-up/sign-in. Requires necessary modules in
the right locations.
