+++ assets/scripts/app.js
@@ -1,0 +1,6 @@
+'use strict'
+
+module.exports = {
+  // host: 'http://tic-tac-toe.wdibos.com'
+  host: 'http://localhost:4741'
+}
