+++ assets/scripts/config.js
@@ -2,7 +2,7 @@
 
 const config = {
   apiOrigins: {
-    production: 'https://ga-wdi-boston.herokuapp.com'
+    production: 'http://localhost:4741'
   }
 }
